+++ internal/prober/multihttp/multihttp.go
@@ -39,7 +39,8 @@ func NewProber(ctx context.Context, check sm.Check, logger zerolog.Logger, runne
 	}
 
 	p.config = settingsToModule(check.Settings.Multihttp)
-	p.config.Timeout = time.Duration(check.Timeout) * time.Millisecond
+	timeout := time.Duration(check.Timeout) * time.Millisecond
+	p.config.Timeout = timeout
 
 	script, err := settingsToScript(check.Settings.Multihttp)
 	if err != nil {