+++ CHANGELOG.md
@@ -2,8 +2,49 @@
 ## [Unreleased]
 
 
+<a name="v0.19.6"></a>
+## [v0.19.6] - 2024-02-06
+### Fix
+- increase max target length
+
+
+<a name="v0.19.5"></a>
+## [v0.19.5] - 2024-02-05
+### Fix
+- check targets must be valid label values
+
+
+<a name="v0.19.4"></a>
+## [v0.19.4] - 2024-01-30
+### Fix
+- allow scripted checks to have anything as the target value ([#592](https://github.com/grafana/synthetic-monitoring-agent/issues/592))
+
+
+<a name="v0.19.3"></a>
+## [v0.19.3] - 2023-12-13
+### Fix
+- test release on PRs
+
+
+<a name="v0.19.2"></a>
+## [v0.19.2] - 2023-12-13
+
+<a name="v0.19.1"></a>
+## [v0.19.1] - 2023-11-20
+
+<a name="v0.19.0"></a>
+## [v0.19.0] - 2023-11-07
+### Feature
+- add k6 to docker image
+
+### Fix
+- make the k6 runner timeout configurable ([#554](https://github.com/grafana/synthetic-monitoring-agent/issues/554))
+- add a `name` label to metrics
+- add k6 binary to release files
+
+
 <a name="v0.18.3"></a>
-## [v0.18.3] - 2023-10-26
+## [v0.18.3] - 2023-10-27
 ### Fix
 - make sure the String() methods match the proto defintion
 
@@ -498,7 +539,14 @@
 <a name="v0.0.1"></a>
 ## v0.0.1 - 2020-06-24
 
-[Unreleased]: https://github.com/grafana/synthetic-monitoring-agent/compare/v0.18.3...HEAD
+[Unreleased]: https://github.com/grafana/synthetic-monitoring-agent/compare/v0.19.6...HEAD
+[v0.19.6]: https://github.com/grafana/synthetic-monitoring-agent/compare/v0.19.5...v0.19.6
+[v0.19.5]: https://github.com/grafana/synthetic-monitoring-agent/compare/v0.19.4...v0.19.5
+[v0.19.4]: https://github.com/grafana/synthetic-monitoring-agent/compare/v0.19.3...v0.19.4
+[v0.19.3]: https://github.com/grafana/synthetic-monitoring-agent/compare/v0.19.2...v0.19.3
+[v0.19.2]: https://github.com/grafana/synthetic-monitoring-agent/compare/v0.19.1...v0.19.2
+[v0.19.1]: https://github.com/grafana/synthetic-monitoring-agent/compare/v0.19.0...v0.19.1
+[v0.19.0]: https://github.com/grafana/synthetic-monitoring-agent/compare/v0.18.3...v0.19.0
 [v0.18.3]: https://github.com/grafana/synthetic-monitoring-agent/compare/v0.18.2...v0.18.3
 [v0.18.2]: https://github.com/grafana/synthetic-monitoring-agent/compare/v0.18.1...v0.18.2
 [v0.18.1]: https://github.com/grafana/synthetic-monitoring-agent/compare/v0.18.0...v0.18.1